Modify upgrade not terminate after lockfile detected
Previously, when the upgrade failed in the initram the file
/sysroot/root/tmp_leapp_py3/.leapp_upgrade_failed has been generated and
upon detecting this file leapp triggered an emergency shell. This caused
the original failure to be hidden from the customer.

With this commit, we no longer crash immediately upon detecting the file
but rather continue and "wait" for the underlying issue and error to
emerge.

RHEL-24148

(cherry picked from commit 1f8b8f3259d7daff63dff0a4d630b36e615e416d)