Default channel to GA is not specified otherwise (#1205)
Originally we tried to map by default repositories from particular channels on the source system to their equivalents on the target system. IOW:

    * eus -> eus
    * aus -> aus
    * e4s -> e4s
    * "ga" -> "ga"
    ...

However, it has been revealed this logic should not apply on minor releases for which these non-ga (premium) repositories do not exist. So doing upgrade e.g. to 8.9, 8.10 , 9.3 for which specific eus, etc.. repositories are not defined lead to 404 error.

Discussing this deeply between stakeholders, it has been decided to drop this logic and target always to "ga" repositories unless the leapp is executed with instructions to choose a different channel (using envars, --channel .. option). To prevent this issue.

It's still possible to require mistakenly e.g. "eus" channel for the target release for which the related repositories are not defined. e.g.:
     > leapp upgrade --channel eus --target 8.10

In such a case, the previous errors (404 Not found) can be hit again. But it will not happen by default. In this case, we expect and request people to understand what they want when they use the option.

@pirat89 : Updated commit msg

jira: RHEL-24720
(cherry picked from commit a4e3906fff5d11e0fb94f5dbe10ed653dc2d0bee)