Add RHEL9 container for building and testing
Adding RHEL 9 containers so we can start to test & build packages
in RHEL 9 environemnt.

Since `python3-virtualenv` package is not available in the RHEL 9 UBI,
a workaround is introduced `make install-deps-fedora`, which is used by
the testing containers. If not able to install the package, `pip install
virtualenv` is used as a fallback as this is working in the image.

The relatively recent Python version used in the container means there
is also a relative recent Pylint verision, which introduces 2+ new
checks.

(cherry picked from commit 953bfb1868595a6100522bfd92f4ea4c47e6be1b)