Fix storage scanner parsing error
Fix storagescanner actor crash when parsing the output of,
e.g., 'pvs -a', which used ':' as a separator and caused errors.
The issue occurred because separator ':' is used to split the outputs of executed commands.
This commit resolves the problem by changing the separator to '|'.

Jira: RHEL-34570
(cherry picked from commit d183370047ab5ef139825dfce7a1b4d6f987092f)