tests: update pytest requirements to Py3.6+ compat
Currently we use frozen python modules due to compatibility with
python2. As we want to keep now still Py2 support due to RHEL 7
for a while, we do not want to simply use the newest versions of
current modules. However, with Python 3.12 we see it's really old
and some of old versions (like pytest) does not work as expected.

For that reason, adding conditional requirements for pytest. I chose
now to stick with pytest 6.2.5 which is compatible with Python3.6+
(due to RHEL 8).

Most likely we will need to revise other requirements too, but I focused
now just on obvious issues.

(cherry picked from commit 14937bc95928fe72aae7b0735d455f82db5a4320)