feat(lvm_autoactivation): add lvm autoactivation
Add LVM autoactivation mechanism to the upgrade initramfs. The core
of the mechanism is based on a special udev rule that is triggered
when a new device is detected. The rule then calls two lvm binaries
(which are also included into the upgrade initrams) to activate
the volume groups and logical volumes.

(cherry picked from commit 04a2ec2574da233a41d32f70eab780b6c305ff31)