Generalize the check of third party RPMs
Originally the redhatsignedrpmcheck actor has been written only
with RHEL systems in mind and the expectation was that the only vendor
as such is Red Hat. However, in case of other distributions that's not
true. The existing report could be confusing for users and could set
wrong expectations.

This patch generalizes the actor with other distributions in mind:
 - the actor is renamed to distributionsignedrpmcheck
 - the report is updated to not mention the distribution/vendor name
   directly

Jira: RHEL-80333

Co-authored-by: Petr Stodulka <pstodulk@redhat.com>
(cherry picked from commit 73b2742151d1ac7a864175d5616b615c2a31fdb1)