feat(command_utils): allow missing minor/major version for CentOS
CentOS versioning consists only of the major version number. Therefore,
in case we are running on CentOS, we do not want to enforce version to
have the format MAJOR.MINOR. Rather, we want to check that the version
is of the form MAJOR. This patch introduces the necessary
infrastructure, allowing easy extensibility of version format checking
for also other distributions. In case we do not know what version
format should a distro use, we default to MINOR.MAJOR.

Jira-ref: RHEL-80334
(cherry picked from commit 097981dd9505d1609b91fbed4e6d28e960161926)