Replace ubi:8 containers with centos:8
The networkmanagerconnectionscanner depends on NetworkManager-nmlib and
python3-gobject. These are not available in the UBI 8 repos and the
tests are skipped.

This patch changes the base container from UBI 8 to Centos 8 which has
the packages available and the tests are not skipped.

(cherry picked from commit 17767238d038d36341181e1702b73681ae432439)