Add --target-version as an alias for the --target option
With the introduction of --target-os option, the --target option is a
bit ambiguous, therefore --target-version is added as an alias.

The leapp-framework dependency version is bumped from 6.1 to 6.2,
because 6.2 introduces the ability to add an alias to an option.

Jira: RHEL-110563
(cherry picked from commit 2df44df07819d8b1e6aaba06534d38fdcc244335)