cs->rhel: Handle potentially unreleased target version
On CS to RHEL upgrades, particularly on 9->10, there is only one upgrade
path defined, CS 9 -> latest RHEL 10 (10.X). However a situation may occur,
in which the latest RHEL version has not yet been publicly released, e.g.
in pre-release builds.

This is problematic because the upgrade fails if the content is not yet
available. If this happens the user is informed via a hint in the error
message to specify the latest available RHEL version (the previous minor
version) manually using the --target-version CLI option.

To make leapp consider the previous minor version as supported without
adding it to upgrade_paths.json (which would affect RHEL->RHEL
upgrades), it is added to the IPUConfig message after processing the
json.

Also, in the error message, move the existing proxy hints from 'details'
to 'hint', otherwise the original exeception message, which is in
'details' is overwritten.

Jira: RHEL-110563
(cherry picked from commit 3c74cd2e237df1646b73238de51bd0d4e019ec82)