Makefile: Do copy commands/ dir in containerized tests
In 7d9ae2c0 the commands dir copying was changed to copy the dir to
tut/lib/$$_VENV/site-packages/leapp/cli/commands/ instead of the normal
location in the container.

This fixed the problem that modifications on the host were not reflected
in the testing containers. However a new problem was introduced -
modifications to tests in the commands directory are not getting
reflected in the testing containers.

This patch fixes that by copying the entire commands directory both to
the normal location and the virtual env (path above).

(cherry picked from commit bf1b5f5f537ff163470b29d8bb7ba452901368eb)