fix unsupported source version crash
This patch fixes a crash when executing upgrade on a system with version
that is not defined in the supported upgrade paths map and when the
fallback mechanism for this particular case fails as well. The upgrade
process is now terminated with an error instead of a trace back.

Jira: RHEL-120252
(cherry picked from commit 2b4d631e975235e2b517fd7a942136e6cc33ed12)