Update EFI entry and directory during conversion
After an upgrade+conversion, the upgraded system might end up with an
EFI boot loader entry from the source system pointing the shim or grub
EFI binary from the source system. E.g. Centos Stream pointing to
\EFI\centos\shimx64.efi instead of Red Hat Enterprise Linux pointing to
\EFI\redhat\shimx64.efi.

This patch introduces a new convert/updateefi actor that adds a new EFI
boot entry for the target system with the correct EFI binary, removes
the source system directory on ESP, e.g /boot/efi/EFI/centos, and
finally removes the source system EFI boot entry (see below).

This is handled for both x86_64 and aarch64 EFI systems. However,
in case of IPU 8 -> 9 on ARM systems it's little bit conflicting with
the existing fixes (the efi_reboot_fix library) which expects already
to set the EFI boot entry for the next boot. We will need to look
at it closer in future but partially it's resolved already.

Additionally, if the source EFI directory is not empty (usually when
it contains custom files), do not remove it and produce a post-upgrade
report instead, instructing the user to review the content and handle
it manually.

This currently does not work with SecureBoot enabled. The problem
probably is, in my understanding, that the shim and grub only contain
keys for verifying signatures of the source distro. Our upgrade boot
entry uses the target distros kernel, which is signed by the target
distro. And so the shim/grub of the source distro rejects it.

Jira: RHEL-133549
(cherry picked from commit d4937406c14b6ea50be54b3ecfe50c8c4919cbb8)